% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafica_T2.R
\name{grafica_T2}
\alias{grafica_T2}
\title{Grafica T2 de Hotelling}
\usage{
grafica_T2(mu, sigma, X, LCS, d, k)
}
\arguments{
\item{mu:}{El vector de las medias de las variables (Parametros del proceso cuando esta bajo control).}

\item{sigma:}{La matriz de varianzas y covarianzas (Parametros del proceso cuando esta bajo control).}

\item{X:}{es la matriz que contiene las variables}

\item{LCS:}{El límite de control superior.}

\item{d:}{La distancia de mahalanobis.}

\item{k:}{El tamaño de la muestra.}
}
\value{
Grafico de control multivariante T2 de Hotelling;

ARL0: Promedio de longitud de rachas cuando el proceso esta bajo control;

ARL1: Promedio de longitud de rachas cuando el proceso esta fuera de control.
}
\description{
Esta funcion permite realizar la grafica de control T2 de Hotelling y obtener sus ARL's bajo y fuera de control.
}
\examples{
###Simulación de datos normales multivariados###
mu=c(4.5,7,8.45)
sigma=matrix(c(2,1.5,2.4,1.5,3,3.1,2.4,3.1,4),ncol=3)
#la matriz de covarianzas debe ser definida positiva
#comprobación por medio de los valores propios de la matriz (>0)
eigen(sigma)
library(MASS)
#datos multivariantes
dmulti=mvrnorm(250,mu,sigma)
colnames(dmulti)=c("va1","va2","va3")
#Datos a utilizar
#n1----->50 dmulti[1:50,]
#n2----->100 dmulti[1:100,]
#n5----->250 dmulti
dmulti_n1=dmulti[1:50,]
dmulti_n2=dmulti[1:100,]
#vector de medias bajo control
mu0=c(5.4,6.8,8.5)
#uso de la funcion T2 de Hotelling 
grafica_T2(mu0,sigma,dmulti_n1,k=1,LCS=14.321,d=1.2)#n=1
grafica_T2(mu0,sigma,dmulti,k=5,LCS=14.321,d=2)#n=5
}
\references{
[1] Aparisi, F.; Epprecht, E.; Ruiz, O.; Veiga A., "Reducing Sampling Costs of Multivariate SPC with a Double-Dimension T2 Control Chart",” International Journal of Production Research, vol. 1, no. 1, pag. 1-15, 2013;

[2] Aparisi, F.; Epprecht, E.; Ruiz, O., "T2 Control Charts with Variable Dimension", Journal of Quality Technology, vol. 44, no. 4, pp. 375-393, 2012;

[3] Ruiz, O., “Gráficos de control de calidad multivariantes con dimension variable” Tesis doctoral, Dept. de Estadistica e Investigación Operativa, Univ. Pol. de Valencia, España, 2013.
}
\keyword{Hotelling}
\keyword{T2}
\keyword{de}

