% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafica_T2DV.R
\name{grafica_T2DV}
\alias{grafica_T2DV}
\title{Grafico de dimension variable VDT2}
\usage{
grafica_T2DV(p1, p, w, CLp1, CLp, disp1, disp, mu, sigma, X, k,
  ProbInic_p1 = 1)
}
\arguments{
\item{p1:}{Numero de variables p1 (Variables fáciles de medir o baratas).}

\item{p2:}{Numero de variables p2 (Variables complejas de medir o costosas).}

\item{p:}{Numero total de variables.}

\item{w:}{Limite de alerta o de advertencia.}

\item{CLP1:}{Limite de control p1.}

\item{CLP:}{Limite de control p.}

\item{disp1:}{Distancia de mahalanobis de p1.}

\item{disp:}{Distancia de mahalanobis de p.}

\item{mu:}{El vector de las medias de las variables (Parametros del proceso cuando esta bajo control).}

\item{sigma:}{La matriz de varianzas y covarianzas (Parametros del proceso cuando esta bajo control).}

\item{X:}{es la matriz que contiene las variables.}

\item{k:}{El tamaño de la muestra.}
}
\value{
Grafico de control multivariante de dimension variable.

ARL0: Promedio de longitud de rachas cuando el proceso esta bajo control.

ARL1: Promedio de longitud de rachas cuando el proceso esta fuera de control.
}
\description{
Esta funcion permite realizar graficos de control de dimension variable y obtener sus ARL's bajo y fuera de control.
}
\examples{
###Simulación de datos normales multivariados###
mu=c(4.5,7,8.45)
sigma=matrix(c(2,1.5,2.4,1.5,3,3.1,2.4,3.1,4),ncol=3)
#la matriz de covarianzas debe ser definida positiva
#comprobación por medio de los valores propios de la matriz (>0)
eigen(sigma)
library(MASS)
#datos multivariantes
dmulti=mvrnorm(250,mu,sigma)
colnames(dmulti)=c("va1","va2","va3")
#Datos a utilizar
#n1----->50 dmulti[1:50,]
#n2----->100 dmulti[1:100,]
#n5----->250 dmulti
dmulti_n1=dmulti[1:50,]
dmulti_n2=dmulti[1:100,]
#vector de medias bajo control
mu0=c(5.4,6.8,8.5)
uso de la función T2 con dimensión variable
grafica_T2DV(p1=2,p=3,w=3.83,CLp1=17.09,CLp=10.62,disp1=1,disp=1.2,mu=mu0,sigma=sigma,X=dmulti_n1,k=1)
grafica_T2DV(p1=2,p=3,w=3.59,CLp1=18.94,CLp=10.82,disp1=1.5,disp=2,mu=mu0,sigma=sigma,X=dmulti,k=5)
#La muestra X debe estar ordenada [p1,p] primero las variables p1: faciles de medir ; y luego las p2: complejas de medir
}
\references{
[1] Aparisi, F.; Epprecht, E.; Ruiz, O.; Veiga A., "Reducing Sampling Costs of Multivariate SPC with a Double-Dimension T2 Control Chart",” International Journal of Production Research, vol. 1, no. 1, pag. 1-15, 2013.

[2] Aparisi, F.; Epprecht, E.; Ruiz, O., "T2 Control Charts with Variable Dimension", Journal of Quality Technology, vol. 44, no. 4, pp. 375-393, 2012.

[3] Ruiz, O., “Gráficos de control de calidad multivariantes con dimensión variable” Tesis doctoral, Dept. de Estadística e Investigacion Operativa, Univ. Pol. de Valencia, Espana, 2013.
}
\keyword{de}
\keyword{dimension}
\keyword{grafico}
\keyword{variable}

